requirements.txt
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clickhouse-connect==0.7.8
Countrydetails==1.0.8
Faker==25.2.0
idna==3.7
kafka-python==2.0.2
lz4==4.3.3
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
zstandard==0.22.0
